(CAT-1901) - Handle 'No URLs in mirrorlist' error for 'appstream' repo in CentOS 8
This commit addresses an issue specific to CentOS 8 where the build process
fails due to an inability to download metadata for the 'appstream' repository.
The error is caused by the absence of URLs in the mirrorlist.

In CentOS 8, the 'appstream' repository is used to provide additional packages
that may be needed by the system. However, due to changes in the way CentOS 8
handles repositories, the mirrorlist for 'appstream' sometimes does not contain
 any URLs, leading to this error.

The Dockerfile has been updated to handle this scenario by modifying the
repository configuration files in '/etc/yum.repos.d/'. If the base image
tag is 'stream8', the script now comments out the 'mirrorlist' line and
replaces it with a 'baseurl' line pointing to 'http://vault.centos.org',
ensuring that the build process can proceed without encountering the
'No URLs in mirrorlist' error.